refactor motif into submodule
