+++ ndspflow/motif/__init__.py
@@ -1,0 +1,1 @@
+from .motif import extract_motifs
